made opensearch data store new main odp data store
