+++ cate/ds/__init__.py
@@ -44,12 +44,14 @@ def cate_init():
     # Plugin initializer.
     # Sets the default data store.
 
-    # from .esa_cci_ftp import set_default_data_store
     from .esa_cci_odp import set_default_data_store
     set_default_data_store()
 
-    from .esa_cci_odp_os import add_data_store
-    add_data_store()
+    # from .esa_cci_ftp import set_default_data_store
+    import os
+    if(os.environ.get('USE_ODP_LEGACY_DATA_STORE', False)):
+        from .esa_cci_odp_legacy import add_data_store
+        add_data_store()
 
     from .local import add_to_data_store_registry
     add_to_data_store_registry()
