+++ cate/ds/esa_cci_odp_legacy.py
@@ -23,68 +23,61 @@
 Description
 ===========
 
-This plugin module adds the ESA CCI Open Data Portal's (ODP) OpenSearch service to
+This plugin module adds the ESA CCI Open Data Portal's (ODP) ESGF service to
 the data store registry.
 
 Verification
 ============
 
 The module's unit-tests are located in
-`test/ds/test_esa_cci_odp_os.py <https://github.com/CCI-Tools/cate/blob/master/test/ds/test_esa_cci_odp_os.py>`_
+`test/ds/test_esa_cci_odp_legacy.py <https://github.com/CCI-Tools/cate/blob/master/test/ds/test_esa_cci_ftp.py>`_
 and may be executed using
-``$ py.test test/ds/test_esa_cci_odp.py --cov=cate/ds/esa_cci_odp_os.py`` for extra code coverage information.
+``$ py.test test/ds/test_esa_cci_odp_legacy.py --cov=cate/ds/esa_cci_odp_legacy.py`` for extra code coverage information.
 
 Components
 ==========
 """
-import aiofiles
-import aiohttp
-import asyncio
 import json
 import logging
 import os
 import re
 import socket
+import ssl
 import urllib.parse
 import urllib.request
 from collections import OrderedDict
 from datetime import datetime, timedelta
-import lxml.etree as etree
-from typing import Sequence, Tuple, Optional, Any, Dict, List, Union
+from math import ceil
+from typing import Sequence, Tuple, Optional, Any, Dict, List
 from urllib.error import URLError, HTTPError
 
 import pandas as pd
 import xarray as xr
+from owslib.csw import CatalogueServiceWeb
+from owslib.namespaces import Namespaces
 
 from cate.conf import get_config_value, get_data_stores_path
 from cate.conf.defaults import NETCDF_COMPRESSION_LEVEL
-from cate.core.ds import DATA_STORE_REGISTRY, NetworkError, DataStore, DataSource, Schema, open_xarray_dataset, \
-    DataStoreNotice
+from cate.core.ds import DATA_STORE_REGISTRY, DataAccessError, NetworkError, DataStore, DataSource, Schema, \
+    open_xarray_dataset, DataStoreNotice
 from cate.core.opimpl import subset_spatial_impl, normalize_impl, adjust_spatial_attrs_impl
 from cate.core.types import PolygonLike, TimeLike, TimeRange, TimeRangeLike, VarNamesLike
 from cate.ds.local import add_to_data_store_registry, LocalDataSource, LocalDataStore
 from cate.util.monitor import Cancellation, Monitor
 
-ESA_CCI_ODP_DATA_STORE_ID = 'esa_cci_odp_os'
+ESA_CCI_ODP_DATA_STORE_ID = 'esa_cci_odp_legacy'
 
 __author__ = "Norman Fomferra (Brockmann Consult GmbH), " \
-             "Tonio Fincke (Brockmann Consult GmbH), " \
              "Marco Zühlke (Brockmann Consult GmbH), " \
              "Chris Bernat (Telespazio VEGA UK Ltd), " \
              "Paolo Pesciullesi (Telespazio VEGA UK Ltd)"
 
-_OPENSEARCH_CEDA_URL = "http://opensearch-test.ceda.ac.uk/opensearch/request"
+_ESGF_CEDA_URL = "https://cci-odp-index.ceda.ac.uk/esg-search/search/"
+_CSW_CEDA_URL = "https://csw-cci.ceda.ac.uk/geonetwork/srv/eng/csw-CEDA-CCI"
 
-ODD_NS = {'os': 'http://a9.com/-/spec/opensearch/1.1/',
-          'param': 'http://a9.com/-/spec/opensearch/extensions/parameters/1.0/'}
-DESC_NS = {'gmd': 'http://www.isotc211.org/2005/gmd',
-           'gml': 'http://www.opengis.net/gml/3.2',
-           'gco': 'http://www.isotc211.org/2005/gco',
-           'gmx': 'http://www.isotc211.org/2005/gmx',
-           'xlink': 'http://www.w3.org/1999/xlink'
-           }
+_TIMESTAMP_FORMAT = "%Y-%m-%d %H:%M:%S"
 
-_TIMESTAMP_FORMAT = "%Y-%m-%dT%H:%M:%S"
+_REFERENCE_DATA_SOURCE_TYPE = "OPEN_DATA_PORTAL"
 
 _RE_TO_DATETIME_FORMATS = patterns = [(re.compile(14 * '\\d'), '%Y%m%d%H%M%S'),
                                       (re.compile(12 * '\\d'), '%Y%m%d%H%M'),
@@ -101,11 +94,17 @@ _TIME_FREQUENCY_TO_TIME_DELTA = dict([
     ('yr', timedelta(days=365)),
 ])
 
-_ODP_PROTOCOL_HTTP = 'Download'
-_ODP_PROTOCOL_OPENDAP = 'Opendap'
+_YEAR_REALIZATION = re.compile(4 * '\\d')
 
-_TIMEOUT = 10
-_MAX_RESULTS = 1000
+_ODP_PROTOCOL_HTTP = 'HTTPServer'
+_ODP_PROTOCOL_OPENDAP = 'OPENDAP'
+
+_ODP_AVAILABLE_PROTOCOLS_LIST = [_ODP_PROTOCOL_HTTP, _ODP_PROTOCOL_OPENDAP]
+
+_CSW_TIMEOUT = 10
+_CSW_MAX_RESULTS = 1000
+_CSW_METADATA_CACHE_FILE = 'catalogue_metadata.xml'
+_CSW_CACHE_FILE = 'catalogue.xml'
 
 _LOG = logging.getLogger('cate')
 
@@ -130,7 +129,7 @@ def add_data_store():
     All data sources of the FTP data store are read from a JSON file ``esa_cci_ftp.json`` contained in this package.
     This JSON file has been generated from a scan of the entire FTP tree.
     """
-    DATA_STORE_REGISTRY.add_data_store(EsaCciOdpOsDataStore())
+    DATA_STORE_REGISTRY.add_data_store(EsaCciOdpLegacyDataStore())
 
 
 def find_datetime_format(filename: str) -> Tuple[Optional[str], int, int]:
@@ -161,224 +160,49 @@ def get_exclude_variables_fix_known_issues(ds_id: str) -> [str]:
         return []
 
 
-async def _extract_metadata_from_odd_url(session = None, odd_url: str = None) -> dict:
-    if session is None:
-        session = aiohttp.ClientSession()
-    if not odd_url:
-        return {}
-    resp = await session.request(method='GET', url=odd_url)
-    resp.raise_for_status()
-    xml_text = await resp.read()
-    return _extract_metadata_from_odd(etree.XML(xml_text))
-
-
-def _extract_metadata_from_odd(odd_xml: etree.XML) -> dict:
-    metadata = {}
-    metadata_names = {'ecv': ['ecv', 'ecvs'], 'frequency': ['time_frequency', 'time_frequencies'],
-                      'institute': ['institute', 'institutes'],
-                      'processingLevel': ['processing_level', 'processing_levels'],
-                      'productString': ['product_string', 'product_strings'],
-                      'productVersion': ['product_version', 'product_versions'],
-                      'dataType': ['data_type', 'data_types'], 'sensor': ['sensor_id', 'sensor_ids'],
-                      'platform': ['platform_id', 'platform_ids'], 'fileFormat': ['file_format', 'file_formats']}
-    for param_elem in odd_xml.findall('os:Url/param:Parameter', namespaces=ODD_NS):
-        if param_elem.attrib['name'] in metadata_names:
-            param_content = _get_from_param_elem(param_elem)
-            if param_content:
-                if type(param_content) == str:
-                    metadata[metadata_names[param_elem.attrib['name']][0]] = param_content
-                else:
-                    metadata[metadata_names[param_elem.attrib['name']][1]] = param_content
-    return metadata
-
-
-def _get_from_param_elem(param_elem: etree.Element) -> Optional[Union[str, List[str]]]:
-    options = param_elem.findall('param:Option', namespaces=ODD_NS)
-    if not options:
-        return None
-    if len(options) == 1:
-        return options[0].get('value')
-    return [option.get('value') for option in options]
-
-
-async def _extract_metadata_from_descxml_url(session = None, descxml_url: str = None) -> dict:
-    if session is None:
-        session = aiohttp.ClientSession()
-    if not descxml_url:
-        return {}
-    resp = await session.request(method = 'GET', url = descxml_url)
-    resp.raise_for_status()
-    descxml = etree.XML(await resp.read())
-    try:
-        return _extract_metadata_from_descxml(descxml)
-    except etree.ParseError:
-        _LOG.info(f'Cannot read metadata from {descxml_url} due to parsing error.')
-        return {}
-
-
-def _extract_metadata_from_descxml(descxml: etree.XML) -> dict:
-    metadata = {}
-    metadata_elems = {
-        'abstract': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:abstract/gco:CharacterString',
-        'title': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:citation/gmd:CI_Citation/gmd:title/'
-                 'gco:CharacterString',
-        'licences': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:resourceConstraints/gmd:MD_Constraints/'
-                    'gmd:useLimitation/gco:CharacterString',
-        'bbox_minx': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:extent/gmd:EX_Extent/'
-                     'gmd:geographicElement/gmd:EX_GeographicBoundingBox/gmd:westBoundLongitude/gco:Decimal',
-        'bbox_miny': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:extent/gmd:EX_Extent/'
-                     'gmd:geographicElement/gmd:EX_GeographicBoundingBox/gmd:southBoundLatitude/gco:Decimal',
-        'bbox_maxx': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:extent/gmd:EX_Extent/'
-                     'gmd:geographicElement/gmd:EX_GeographicBoundingBox/gmd:eastBoundLongitude/gco:Decimal',
-        'bbox_maxy': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:extent/gmd:EX_Extent/'
-                     'gmd:geographicElement/gmd:EX_GeographicBoundingBox/gmd:northBoundLatitude/gco:Decimal',
-        'temporal_coverage_start': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:extent/gmd:EX_Extent/'
-                                   'gmd:temporalElement/gmd:EX_TemporalExtent/gmd:extent/gml:TimePeriod/'
-                                   'gml:beginPosition',
-        'temporal_coverage_end': 'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:extent/gmd:EX_Extent/'
-                                 'gmd:temporalElement/gmd:EX_TemporalExtent/gmd:extent/gml:TimePeriod/gml:endPosition'
-    }
-    for identifier in metadata_elems:
-        content = _get_element_content(descxml, metadata_elems[identifier])
-        if content:
-            metadata[identifier] = content
-    metadata_elems_with_replacement = {'file_formats': [
-        'gmd:identificationInfo/gmd:MD_DataIdentification/gmd:resourceFormat/gmd:MD_Format/gmd:name/'
-        'gco:CharacterString', 'Data are in NetCDF format', '.nc']
-    }
-    for metadata_elem in metadata_elems_with_replacement:
-        content = _get_replaced_content_from_descxml_elem(descxml, metadata_elems_with_replacement[metadata_elem])
-        if content:
-            metadata[metadata_elem] = content
-    metadata_linked_elems = {
-        'publication_date': ['gmd:identificationInfo/gmd:MD_DataIdentification/gmd:citation/'
-                             'gmd:CI_Citation/gmd:date/gmd:CI_Date/gmd:dateType/gmd:CI_DateTypeCode', 'publication',
-                             '../../gmd:date/gco:DateTime'],
-        'creation_date': ['gmd:identificationInfo/gmd:MD_DataIdentification/gmd:citation/'
-                          'gmd:CI_Citation/gmd:date/gmd:CI_Date/gmd:dateType/gmd:CI_DateTypeCode', 'creation',
-                          '../../gmd:date/gco:DateTime']
-    }
-    for identifier in metadata_linked_elems:
-        content = _get_linked_content_from_descxml_elem(descxml, metadata_linked_elems[identifier])
-        if content:
-            metadata[identifier] = content
-
-    anchor_elem_path = 'gmd:identificationInfo/gmd:MD_DataIdentification/' \
-                       'gmd:descriptiveKeywords/gmd:MD_Keywords/gmd:keyword/gmx:Anchor'
-    anchor_collect_elems = {
-        'http://vocab-test.ceda.ac.uk/collection/cci/platform/': ['platform_id', 'platform_ids'],
-        'http://vocab-test.ceda.ac.uk/collection/cci/sensor/': ['sensor_id', 'sensor_ids'],
-        'http://vocab-test.ceda.ac.uk/collection/cci/procLev/': ['processing_level', 'processing_levels'],
-        'http://vocab-test.ceda.ac.uk/collection/cci/freq/': ['time_frequency', 'time_frequencies']
-    }
-    metadata_collect = {}
-    for anchor_collect_elem in anchor_collect_elems:
-        metadata_collect[anchor_collect_elem] = []
-    anchor_elems = descxml.findall(anchor_elem_path, namespaces=DESC_NS)
-    for anchor_elem in anchor_elems:
-        href_attrib = anchor_elem.xpath('@xlink:href', namespaces=DESC_NS)
-        if href_attrib:
-            for identifier in anchor_collect_elems:
-                if str(href_attrib[0]).startswith(identifier):
-                    metadata_collect[identifier].append(anchor_elem.text)
-    for identifier in metadata_collect:
-        if len(metadata_collect[identifier]) == 1:
-            metadata[anchor_collect_elems[identifier][0]] = metadata_collect[identifier][0]
-        elif len(metadata_collect[identifier]) > 0:
-            metadata[anchor_collect_elems[identifier][1]] = metadata_collect[identifier]
-    return metadata
-
-
-def _get_element_content(descxml: etree.XML, path: str) -> Optional[Union[str, List[str]]]:
-    elems = descxml.findall(path, namespaces=DESC_NS)
-    if not elems:
-        return None
-    if len(elems) == 1:
-        return elems[0].text
-    return [elem.text for elem in elems]
-
-
-def _get_replaced_content_from_descxml_elem(descxml: etree.XML, paths: List[str]) -> Optional[str]:
-    descxml_elem = descxml.find(paths[0], namespaces=DESC_NS)
-    if descxml_elem is None:
-        return None
-    if descxml_elem.text == paths[1]:
-        return paths[2]
-
-
-def _get_linked_content_from_descxml_elem(descxml: etree.XML, paths: List[str]) -> Optional[str]:
-    descxml_elems = descxml.findall(paths[0], namespaces=DESC_NS)
-    if descxml is None:
-        return None
-    for descxml_elem in descxml_elems:
-        if descxml_elem.text == paths[1]:
-            return _get_element_content(descxml_elem, paths[2])
-
-
-async def _fetch_feature_at(session, base_url, query_args, index) -> Optional[Dict]:
-    paging_query_args = dict(query_args or {})
-    maximum_records = 1
-    paging_query_args.update(startPage=index, maximumRecords=maximum_records, httpAccept='application/geo+json',
-                            fileFormat='.nc')
-    url = base_url + '?' + urllib.parse.urlencode(paging_query_args)
-    resp = await session.request(method='GET', url=url)
-    resp.raise_for_status()
-    json_text = await resp.read()
-    json_dict = json.loads(json_text.decode('utf-8'))
-    feature_list = json_dict.get("features", [])
-    if len(feature_list) > 0:
-        return feature_list[0]
-    return None
-
-
-async def _fetch_opensearch_feature_list(base_url, query_args, monitor: Monitor = Monitor.NONE) -> List:
+def _fetch_solr_json(base_url, query_args, offset=0, limit=3500, timeout=10, monitor: Monitor = Monitor.NONE):
     """
-    Return JSON value read from Opensearch web service.
-    :return:
+    Return JSON value read from paginated Solr web-service.
     """
-    start_page = 1
-    maximum_records = 1000
-    full_feature_list = []
+    combined_json_dict = None
+    num_found = -1
+    # we don't know ahead of time how many requests are necessary
     with monitor.starting("Loading", 10):
-        async with aiohttp.ClientSession() as session:
-            #todo remove this when the opensearch server provides results of more than 10,000 features
-            while start_page < 11:
-                monitor.progress(work=1)
-                paging_query_args = dict(query_args or {})
-                paging_query_args.update(startPage=start_page, maximumRecords=maximum_records,
-                                         httpAccept='application/geo+json')
-                url = base_url + '?' + urllib.parse.urlencode(paging_query_args)
-                resp = await session.request(method='GET', url=url)
-                resp.raise_for_status()
-                json_text = await resp.read()
-                json_dict = json.loads(json_text.decode('utf-8'))
-                feature_list = json_dict.get("features", [])
-                full_feature_list.extend(feature_list)
-                if len(feature_list) < maximum_records:
-                    break
-                start_page += 1
-    return full_feature_list
+        while True:
+            monitor.progress(work=1)
+            paging_query_args = dict(query_args or {})
+            # noinspection PyArgumentList
+            paging_query_args.update(offset=offset, limit=limit, format='application/solr+json')
+            url = base_url + '?' + urllib.parse.urlencode(paging_query_args)
+            error_message = f"Failed accessing CCI ODP service {base_url}"
+            try:
+                # noinspection PyProtectedMember
+                context = ssl._create_unverified_context()
+                with urllib.request.urlopen(url, timeout=timeout, context=context) as response:
+                    json_text = response.read()
+                    json_dict = json.loads(json_text.decode('utf-8'))
+                    if num_found is -1:
+                        num_found = json_dict.get('response', {}).get('numFound', 0)
+                    if not combined_json_dict:
+                        combined_json_dict = json_dict
+                        if num_found < limit:
+                            break
+                    else:
+                        docs = json_dict.get('response', {}).get('docs', [])
+                        combined_json_dict.get('response', {}).get('docs', []).extend(docs)
+                        if num_found < offset + limit:
+                            break
+            except urllib.error.HTTPError as e:
+                raise DataAccessError(f"{error_message}: {e}") from e
+            except (urllib.error.URLError, socket.timeout) as e:
+                raise NetworkError(f"{error_message}: {e}") from e
+            except OSError as e:
+                raise DataAccessError(f"{error_message}: {e}") from e
+            offset += limit
+    return combined_json_dict
 
 
-def _harmonize_info_field_names(catalogue: dict, single_field_name: str, multiple_fields_name: str,
-                                multiple_items_name: Optional[str] = None):
-    if single_field_name in catalogue and multiple_fields_name in catalogue:
-        if len(multiple_fields_name) == 0:
-            catalogue.pop(multiple_fields_name)
-        elif len(catalogue[multiple_fields_name]) == 1:
-            if catalogue[multiple_fields_name][0] is catalogue[single_field_name]:
-                catalogue.pop(multiple_fields_name)
-            else:
-                catalogue[multiple_fields_name].append(catalogue[single_field_name])
-                catalogue.pop(single_field_name)
-        else:
-            if catalogue[single_field_name] not in catalogue[multiple_fields_name] \
-                    and (multiple_items_name is None or catalogue[single_field_name] != multiple_items_name):
-                catalogue[multiple_fields_name].append(catalogue[single_field_name])
-            catalogue.pop(single_field_name)
-
-
-async def _load_or_fetch_json(fetch_json_function,
+def _load_or_fetch_json(fetch_json_function,
                         fetch_json_args: list = None,
                         fetch_json_kwargs: dict = None,
                         cache_used: bool = False,
@@ -423,15 +247,15 @@ async def _load_or_fetch_json(fetch_json_function,
         # noinspection PyArgumentList
         try:
             # noinspection PyArgumentList
-            json_obj = await fetch_json_function(*(fetch_json_args or []), **(fetch_json_kwargs or {}))
+            json_obj = fetch_json_function(*(fetch_json_args or []), **(fetch_json_kwargs or {}))
             if cache_used:
                 os.makedirs(cache_dir, exist_ok=True)
                 # noinspection PyUnboundLocalVariable
-                async with aiofiles.open(cache_json_file, "w") as fp:
-                    await fp.write(json.dumps(json_obj, indent='  '))
+                with open(cache_json_file, 'w') as fp:
+                    fp.write(json.dumps(json_obj, indent='  '))
                 # noinspection PyUnboundLocalVariable
-                async with aiofiles.open(cache_timestamp_file, "w") as fp:
-                    await fp.write(datetime.utcnow().strftime(_TIMESTAMP_FORMAT))
+                with open(cache_timestamp_file, 'w') as fp:
+                    fp.write(datetime.utcnow().strftime(_TIMESTAMP_FORMAT))
         except Exception as e:
             if cache_json_file and os.path.exists(cache_json_file):
                 with open(cache_json_file) as fp:
@@ -443,164 +267,61 @@ async def _load_or_fetch_json(fetch_json_function,
     return json_obj
 
 
-async def _fetch_data_source_list_json(base_url, query_args, monitor: Monitor = Monitor.NONE) -> Sequence:
-    feature_collection_list = await _fetch_opensearch_feature_list(base_url, query_args, monitor=monitor)
-    catalogue = {}
-    for fc in feature_collection_list:
-        fc_props = fc.get("properties", {})
-        fc_id = fc_props.get("identifier", None)
-        if not fc_id:
-            continue
-        data_source_id = f'esacci.{fc_id}'
-        catalogue[data_source_id] = {}
-        catalogue[data_source_id]['title'] = fc_props.get("title", "")
-        variables = _get_variables_from_feature(fc)
-        catalogue[data_source_id]['variables'] = variables
-        fc_props_links = fc_props.get("links", None)
-        if fc_props_links:
-            search = fc_props_links.get("search", None)
-            if search:
-                odd_url = search[0].get('href', None)
-                if odd_url:
-                    catalogue[data_source_id]['odd_url'] = odd_url
-            described_by = fc_props_links.get("describedby", None)
-            if described_by:
-                metadata_url = described_by[0].get("href", None)
-                if metadata_url:
-                    catalogue[data_source_id]['metadata_url'] = metadata_url
-            # todo consider including this at time of data store reading
-            # catalogue[data_source_id]['metadata'] = await _fetch_meta_info(fc_id, odd_url, metadata_url, variables,
-            # False)
-        _LOG.info(f'Added data source {data_source_id}')
-    return catalogue
+def _fetch_file_list_json(dataset_id: str, dataset_query_id: str, monitor: Monitor = Monitor.NONE):
+    file_index_json_dict = _fetch_solr_json(_ESGF_CEDA_URL,
+                                            dict(type='File',
+                                                 fields='url,title,size',
+                                                 dataset_id=dataset_query_id,
+                                                 replica='false',
+                                                 latest='True',
+                                                 project='esacci'),
+                                            monitor=monitor)
 
+    if not isinstance(file_index_json_dict, dict):
+        return None
 
-def _get_variables_from_feature(feature: dict) -> List:
-    feature_props = feature.get("properties", {})
-    variables = feature_props.get("variables", [])
-    variable_dicts = []
-    for variable in variables:
-        variable_dict = {
-            'name': variable.get("var_id", None),
-            'units': variable.get("units", ""),
-            'long_name': variable.get("long_name", None)}
-        variable_dicts.append(variable_dict)
-    return variable_dicts
-
-
-async def _get_infos_from_feature(session, feature: dict) -> tuple:
-    feature_info = _extract_feature_info(feature)
-    opendap_dds_url = f"{feature_info[4]['Opendap']}.dds"
-    resp = await session.request(method='GET', url=opendap_dds_url)
-    if resp.status >= 400:
-        resp.release()
-        _LOG.warning(f"Could not open {opendap_dds_url}: {resp.status}")
-        return {}, {}
-    content = await resp.read()
-    return _retrieve_infos_from_dds(str(content, 'utf-8').split('\n'))
-
-
-def _retrieve_infos_from_dds(dds_lines: List) -> tuple:
-    dimensions = {}
-    variable_infos = {}
-    dim_info_pattern = '[a-zA-Z0-9_]+ [a-zA-Z0-9_]+[\[\w* = \d{1,7}\]]*;'
-    type_and_name_pattern = '[a-zA-Z0-9_]+'
-    dimension_pattern = '\[[a-zA-Z]* = \d{1,7}\]'
-    for dds_line in dds_lines:
-        if type(dds_line) is bytes:
-            dds_line = str(dds_line, 'utf-8')
-        dim_info_search_res = re.search(dim_info_pattern, dds_line)
-        if dim_info_search_res is None:
-            continue
-        type_and_name = re.findall(type_and_name_pattern, dim_info_search_res.string)
-        if type_and_name[1] not in variable_infos:
-            dimension_names = []
-            variable_dimensions = re.findall(dimension_pattern, dim_info_search_res.string)
-            for variable_dimension in variable_dimensions:
-                dimension_name, dimension_size = variable_dimension[1:-1].split(' = ')
-                dimension_names.append(dimension_name)
-                if dimension_name not in dimensions:
-                    dimensions[dimension_name] = int(dimension_size)
-            variable_infos[type_and_name[1]] = {'data_type': type_and_name[0], 'dimensions': dimension_names}
-    return dimensions, variable_infos
-
-
-async def _fetch_meta_info(dataset_id: str, odd_url: str, metadata_url: str, variables: List, read_dimensions: bool) \
-        -> Dict:
-    async with aiohttp.ClientSession() as session:
-        meta_info_dict = {}
-        if odd_url:
-            meta_info_dict = await _extract_metadata_from_odd_url(session, odd_url)
-        if metadata_url:
-            desc_metadata = await _extract_metadata_from_descxml_url(session, metadata_url)
-            for item in desc_metadata:
-                if item in meta_info_dict and type(meta_info_dict[item]) == list:
-                    meta_info_dict[item].extend(desc_metadata[item])
-                    meta_info_dict[item] = list(dict.fromkeys(meta_info_dict[item]))
-                else:
-                    meta_info_dict[item] = desc_metadata[item]
-        meta_info_dict['dimensions'] = {}
-        meta_info_dict['variable_infos'] = {}
-        if read_dimensions and len(variables) > 0:
-            feature = await _fetch_feature_at(session, _OPENSEARCH_CEDA_URL, dict(parentIdentifier=dataset_id), 1)
-            if feature is not None:
-                feature_dimensions, feature_variable_infos = await _get_infos_from_feature(session, feature)
-                meta_info_dict['dimensions'] = feature_dimensions
-                meta_info_dict['variable_infos'] = feature_variable_infos
-        _harmonize_info_field_names(meta_info_dict, 'file_format', 'file_formats')
-        _harmonize_info_field_names(meta_info_dict, 'platform_id', 'platform_ids', 'multi-platform')
-        _harmonize_info_field_names(meta_info_dict, 'sensor_id', 'sensor_ids', 'multi-sensor')
-        _harmonize_info_field_names(meta_info_dict, 'processing_level', 'processing_levels')
-        _harmonize_info_field_names(meta_info_dict, 'time_frequency', 'time_frequencies')
-        return meta_info_dict
-
-
-async def _fetch_file_list_json(dataset_id: str, monitor: Monitor = Monitor.NONE) -> Sequence:
-    feature_list = await _fetch_opensearch_feature_list(_OPENSEARCH_CEDA_URL,
-                                                        dict(parentIdentifier=dataset_id,
-                                                             fileFormat='.nc'),
-                                                        monitor=monitor)
     file_list = []
-    for feature in feature_list:
-        feature_info = _extract_feature_info(feature)
-        if feature_info[0] in file_list:
-            raise ValueError('filename {} already seen in dataset {}'.format(feature_info[0], dataset_id))
-        file_list.append(feature_info)
-    max_time = datetime.strftime(datetime.max, _TIMESTAMP_FORMAT)
+    docs = file_index_json_dict.get('response', {}).get('docs', [])
+    time_info = None
+    for doc in docs:
+        urls = dict()
+        url_rec_list = doc.get('url', [])
+        for url_rec in url_rec_list:
+            url, mime_type, url_type = url_rec.split('|')
+            urls[url_type] = url
+
+        filename = doc.get('title', None)
+        file_size = doc.get('size', -1)
+        if not filename:
+            filename = os.path.basename(urllib.parse.urlparse(urls[_ODP_PROTOCOL_HTTP])[2])
+        if filename in file_list:
+            raise ValueError('filename {} already seen in dataset {}'
+                             .format(filename, dataset_id))
+        if not time_info:
+            time_info = find_datetime_format(filename)
+        # Start time will be extracted from filename using time_info
+        start_time = None
+        # We also reserve an end_time field, just in case.
+        end_time = None
+        if time_info:
+            time_format, p1, p2 = time_info
+            if time_format:
+                start_time = datetime.strptime(filename[p1:p2], time_format)
+                # Convert back to text, so we can JSON-encode it
+                start_time = datetime.strftime(start_time, _TIMESTAMP_FORMAT)
+        file_list.append([filename, start_time, end_time, file_size, urls])
+
     def pick_start_time(file_info_rec):
-        return file_info_rec[1] if file_info_rec[1] else max_time
+        return file_info_rec[1] if file_info_rec[1] else datetime.max
 
     return sorted(file_list, key=pick_start_time)
 
 
-def _extract_feature_info(feature: dict) -> List:
-    feature_props = feature.get("properties", {})
-    filename = feature_props.get("title", "")
-    date = feature_props.get("date", None)
-    start_time = ""
-    end_time = ""
-    if date and "/" in date:
-        start_time, end_time = date.split("/")
-    elif filename:
-        time_format, p1, p2 = find_datetime_format(filename)
-        if time_format:
-            start_time = datetime.strptime(filename[p1:p2], time_format)
-            # Convert back to text, so we can JSON-encode it
-            start_time = datetime.strftime(start_time, _TIMESTAMP_FORMAT)
-            end_time = start_time
-    file_size = feature_props.get("filesize", 0)
-    related_links = feature_props.get("links", {}).get("related", [])
-    urls = {}
-    for related_link in related_links:
-        urls[related_link.get("title")] = related_link.get("href")
-    return [filename, start_time, end_time, file_size, urls]
-
-
-class EsaCciOdpOsDataStore(DataStore):
+class EsaCciOdpLegacyDataStore(DataStore):
     # noinspection PyShadowingBuiltins
     def __init__(self,
-                 id: str = 'esa_cci_odp_os',
-                 title: str = 'ESA CCI Open Data Portal',
+                 id: str = 'esa_cci_odp_legacy',
+                 title: str = 'ESA CCI Open Data Portal Legacy',
                  index_cache_used: bool = True,
                  index_cache_expiration_days: float = 1.0,
                  index_cache_json_dict: dict = None,
@@ -608,9 +329,11 @@ class EsaCciOdpOsDataStore(DataStore):
         super().__init__(id, title=title, is_local=False)
         self._index_cache_used = index_cache_used
         self._index_cache_expiration_days = index_cache_expiration_days
-        self._catalogue = index_cache_json_dict
+        self._esgf_data = index_cache_json_dict
         self._index_cache_update_tag = index_cache_update_tag
         self._data_sources = []
+
+        self._csw_data = None
 
     @property
     def description(self) -> Optional[str]:
@@ -621,7 +344,9 @@ class EsaCciOdpOsDataStore(DataStore):
         """
         return ("This data store represents the [ESA CCI Open Data Portal](http://cci.esa.int/data)"
                 " in the CCI Toolbox.\n"
-                "It currently provides all CCI data that are published through an Opensearch Interface. "
+                "It currently provides all CCI data that are published through the "
+                "[ESGF-CEDA services](https://esgf-index1.ceda.ac.uk/search/esacci-ceda/) "
+                "(gridded data stored as NetCDF files). "
                 "The store will be extended shortly to also provide TIFF and Shapefile Data, see usage "
                 "notes.\n"
                 "Remote data downloaded to your computer is made available through the *Local Data Store*.")
@@ -686,7 +411,7 @@ class EsaCciOdpOsDataStore(DataStore):
 
     def query(self, ds_id: str = None, query_expr: str = None, monitor: Monitor = Monitor.NONE) \
             -> Sequence['DataSource']:
-        asyncio.run(self._init_data_sources())
+        self._init_data_sources()
         if ds_id or query_expr:
             return [ds for ds in self._data_sources if ds.matches(ds_id=ds_id, query_expr=query_expr)]
         return self._data_sources
@@ -742,25 +467,34 @@ class EsaCciOdpOsDataStore(DataStore):
     def __repr__(self) -> str:
         return "EsaCciOdpDataStore (%s)" % self.id
 
-    async def _init_data_sources(self):
+    def _init_data_sources(self):
         os.makedirs(get_metadata_store_path(), exist_ok=True)
         if self._data_sources:
             return
-        if self._catalogue is None:
-            await self._load_index()
-        if self._catalogue:
-            self._data_sources = []
-            tasks = []
-            for catalogue_item in self._catalogue:
-                tasks.append(self._create_data_source(self._catalogue[catalogue_item], catalogue_item))
-            await asyncio.gather(*tasks)
+        if self._esgf_data is None:
+            self._load_index()
+        if self._esgf_data is None:
+            return
+
+        docs = self._esgf_data.get('response', {}).get('docs', [])
+        data_sources = []
+        if self._csw_data:
+            for catalogue_data in self._csw_data.values():
+                catalogue_item = catalogue_data.copy()
+                catalogue_item.pop('data_sources')
+                for ds_name in catalogue_data.get('data_sources'):
+                    for idx, doc in enumerate(docs):
+                        instance_id = doc.get('instance_id', None)
+                        if ds_name == instance_id:
+                            data_sources.append(EsaCciOdpLegacyDataSource(self, doc, catalogue_item))
+                            del docs[idx]
+                            break
+        else:
+            for doc in docs:
+                data_sources.append(EsaCciOdpLegacyDataSource(self, doc))
+        self._data_sources = data_sources
         self._check_source_diff()
         self._freeze_source()
-
-    async def _create_data_source(self, json_dict: dict, datasource_id: str):
-        data_source = EsaCciOdpOsDataSource(self, json_dict, datasource_id)
-        await data_source.ensure_meta_info_set()
-        self._data_sources.append(data_source)
 
     def _get_update_tag(self):
         """
@@ -796,7 +530,7 @@ class EsaCciOdpOsDataStore(DataStore):
             with open(frozen_file, 'r') as json_in:
                 frozen_source = json.load(json_in)
 
-            ds_new = set([ds.id for ds in self._data_sources])
+            ds_new = set([ds.to_json()['id'] for ds in self._data_sources])
             ds_old = set([ds for ds in frozen_source['data']])
             for ds in (ds_old - ds_new):
                 deleted.append(ds)
@@ -834,105 +568,138 @@ class EsaCciOdpOsDataStore(DataStore):
             save_it = (now > source_ref_time + timedelta(days=1))
 
         if save_it:
-            data = [ds.id for ds in self._data_sources]
+            data = [ds.to_json()['id'] for ds in self._data_sources]
             freezed_source = {'source_ref_time': str(now),
                               'data': data}
             with open(frozen_file, 'w+') as json_out:
                 json.dump(freezed_source, json_out)
 
-    async def _load_index(self):
-        self._catalogue = await _load_or_fetch_json(_fetch_data_source_list_json,
-                                              fetch_json_args=[
-                                                  _OPENSEARCH_CEDA_URL,
-                                                  dict(parentIdentifier='cci')
-                                              ],
-                                              cache_used=self._index_cache_used,
-                                              cache_dir=get_metadata_store_path(),
-                                              cache_json_filename='dataset-list.json',
-                                              cache_timestamp_filename='dataset-list-timestamp.json',
-                                              cache_expiration_days=self._index_cache_expiration_days
-                                              )
+    def _load_index(self):
+        esgf_json_dict = _load_or_fetch_json(_fetch_solr_json,
+                                             fetch_json_args=[
+                                                 _ESGF_CEDA_URL,
+                                                 dict(type='Dataset',
+                                                      replica='false',
+                                                      latest='true',
+                                                      project='esacci')],
+                                             cache_used=self._index_cache_used,
+                                             cache_dir=get_metadata_store_path(),
+                                             cache_json_filename='dataset-list.json',
+                                             cache_timestamp_filename='dataset-list-timestamp.json',
+                                             cache_expiration_days=self._index_cache_expiration_days)
+
+        cci_catalogue_service = EsaCciCatalogueService(_CSW_CEDA_URL)
+        csw_json_dict = _load_or_fetch_json(cci_catalogue_service.getrecords,
+                                            fetch_json_args=[],
+                                            cache_used=self._index_cache_used,
+                                            cache_dir=get_metadata_store_path(),
+                                            cache_json_filename='catalogue.json',
+                                            cache_timestamp_filename='catalogue-timestamp.json',
+                                            cache_expiration_days=self._index_cache_expiration_days)
+
+        self._csw_data = csw_json_dict
+        self._esgf_data = esgf_json_dict
 
 
-INFO_FIELD_NAMES = sorted(["title",
+INFO_FIELD_NAMES = sorted(["realization",
+                           "project",
+                           "number_of_aggregations",
+                           "size",
+                           "product_string",
+                           "platform_id",
+                           "number_of_files",
+                           "product_version",
+                           "time_frequency",
+                           "processing_level",
+                           "sensor_id",
+                           "version",
+                           "cci_project",
+                           "data_type",
+                           # catalogue data fields
                            "abstract",
-                           "licences",
                            "bbox_minx",
                            "bbox_miny",
                            "bbox_maxx",
                            "bbox_maxy",
-                           "temporal_coverage_start",
-                           "temporal_coverage_end",
-                           "file_format",
-                           "file_formats",
-                           "publication_date",
                            "creation_date",
-                           "platform_ids",
-                           "platform_id",
-                           "sensor_ids",
-                           "sensor_id",
-                           "processing_levels",
-                           "processing_level",
-                           "time_frequencies",
-                           "time_frequency",
-                           "ecv",
-                           "institute",
-                           "institutes",
-                           "product_string",
-                           "product_strings",
-                           "product_version",
-                           "product_versions",
-                           "data_type",
-                           "data_types"
+                           "publication_date",
+                           "title",
+                           "data_sources",
+                           "licences",
+                           "temporal_coverage_start",
+                           "temporal_coverage_end"
                            ])
 
 
-class EsaCciOdpOsDataSource(DataSource):
+class EsaCciOdpLegacyDataSource(DataSource):
     def __init__(self,
-                 data_store: EsaCciOdpOsDataStore,
+                 data_store: EsaCciOdpLegacyDataStore,
                  json_dict: dict,
-                 datasource_id: str,
+                 cci_catalogue_data: dict = None,
                  schema: Schema = None):
-        super(EsaCciOdpOsDataSource, self).__init__()
-        self._datasource_id = datasource_id
-        self._raw_id = datasource_id.split('.')[-1]
+        super(EsaCciOdpLegacyDataSource, self).__init__()
+        self._master_id = json_dict.get('master_id', None)
+        self._dataset_id = json_dict.get('id', None)
+        self._instance_id = json_dict.get('instance_id', None)
+
+        if json_dict.get('xlink', None):
+            self._uuid = json_dict.get('xlink')[0].split('|', 1)[0].rsplit('/', 1)[1]
+        else:
+            self._uuid = None
+
         self._data_store = data_store
         self._json_dict = json_dict
-        self._meta_info_dict = None
-        if 'metadata' in self._json_dict:
-            self._meta_info_dict = self._json_dict['metadata']
         self._schema = schema
+        self._catalogue_data = cci_catalogue_data
+
         self._file_list = None
-        self._meta_info = None
+
         self._temporal_coverage = None
+        self._protocol_list = None
+        self._meta_info = None
 
     @property
     def id(self) -> str:
-        return self._datasource_id
+        return self._master_id
 
     @property
-    def data_store(self) -> EsaCciOdpOsDataStore:
+    def uuid(self) -> Optional[str]:
+        return self._uuid
+
+    @property
+    def data_store(self) -> EsaCciOdpLegacyDataStore:
         return self._data_store
 
     @property
     def spatial_coverage(self) -> Optional[PolygonLike]:
-        if self._meta_info_dict \
-                and self._meta_info_dict.get('bbox_minx', None) and self._meta_info_dict.get('bbox_miny', None) \
-                and self._meta_info_dict.get('bbox_maxx', None) and self._meta_info_dict.get('bbox_maxy', None):
+        if self._catalogue_data \
+                and self._catalogue_data.get('bbox_minx', None) and self._catalogue_data.get('bbox_miny', None) \
+                and self._catalogue_data.get('bbox_maxx', None) and self._catalogue_data.get('bbox_maxy', None):
             return PolygonLike.convert([
-                self._meta_info_dict.get('bbox_minx'),
-                self._meta_info_dict.get('bbox_miny'),
-                self._meta_info_dict.get('bbox_maxx'),
-                self._meta_info_dict.get('bbox_maxy')
+                self._catalogue_data.get('bbox_minx'),
+                self._catalogue_data.get('bbox_miny'),
+                self._catalogue_data.get('bbox_maxx'),
+                self._catalogue_data.get('bbox_maxy')
             ])
         return None
 
     def temporal_coverage(self, monitor: Monitor = Monitor.NONE) -> Optional[TimeRange]:
         if not self._temporal_coverage:
-            temp_coverage_start = self._meta_info_dict.get('temporal_coverage_start', None)
-            temp_coverage_end = self._meta_info_dict.get('temporal_coverage_end', None)
+            temp_coverage_start = self._catalogue_data.get('temporal_coverage_start', None)
+            temp_coverage_end = self._catalogue_data.get('temporal_coverage_end', None)
             if temp_coverage_start and temp_coverage_end:
                 self._temporal_coverage = TimeRangeLike.convert("{},{}".format(temp_coverage_start, temp_coverage_end))
+                # ODP datasets that are split into per-year datasets
+                # have the year they are covering in the 'realization' attribute
+                # the CSW does not have separate temporal coverages for them
+                realization = self._json_dict.get('realization', None)
+                if realization and len(realization):
+                    matcher = _YEAR_REALIZATION.match(realization[0])
+                    if matcher:
+                        year = matcher.group(0)
+                        rel_start = max(self._temporal_coverage[0], datetime(int(year), 1, 1))
+                        rel_end = min(self._temporal_coverage[1], datetime(int(year) + 1, 1, 1) - timedelta(seconds=1))
+                        self._temporal_coverage = (rel_start, rel_end)
             else:
                 self.update_file_list(monitor)
         if self._temporal_coverage:
@@ -941,39 +708,33 @@ class EsaCciOdpOsDataSource(DataSource):
 
     @property
     def variables_info(self):
-        variables = []
-        coordinate_variable_names = ['lat', 'lon', 'time', 'lat_bnds', 'lon_bnds', 'time_bnds', 'crs']
-        for variable in self._json_dict['variables']:
-            if 'variable_infos' in self._meta_info and variable['name'] in self._meta_info['variable_infos'] and \
-                    len(self._meta_info['variable_infos'][variable['name']]['dimensions']) == 0:
-                continue
-            if 'dimensions' in self._meta_info and variable['name'] in self._meta_info['dimensions']:
-                continue
-            if variable['name'] not in coordinate_variable_names:
-                variables.append(variable)
-        return variables
+        return self._variables_list()
 
     @property
     def schema(self) -> Schema:
         return self._schema
 
     @property
-    def title(self) -> Optional[str]:
-        return self._json_dict['title']
-
-    @property
     def meta_info(self) -> OrderedDict:
         # noinspection PyBroadException
         if not self._meta_info:
-            self._meta_info = OrderedDict()
+            meta_info = OrderedDict()
             for name in INFO_FIELD_NAMES:
-                value = self._meta_info_dict.get(name, None)
+                value = self._json_dict.get(name, None)
                 # Many values in the index JSON are one-element lists: turn them into scalars
                 if isinstance(value, list) and len(value) == 1:
                     value = value[0]
-                if value is not None:
-                    self._meta_info[name] = value
-            self._meta_info['variables'] = self.variables_info
+                meta_info[name] = value
+
+            meta_info['protocols'] = self.protocols
+            meta_info['variables'] = self._variables_list()
+            meta_info['uuid'] = self._uuid
+
+            if self._catalogue_data:
+                meta_info.update(self._catalogue_data)
+
+            self._meta_info = meta_info
+
         return self._meta_info
 
     @property
@@ -997,19 +758,48 @@ class EsaCciOdpOsDataSource(DataSource):
                         coverage[date_from] = date_to
         return coverage
 
+    def _variables_list(self):
+        variable_names = self._json_dict.get('variable', [])
+        default_list = len(variable_names) * [None]
+        units = self._json_dict.get('variable_units', default_list)
+        long_names = self._json_dict.get('variable_long_name', default_list)
+        standard_names = self._json_dict.get('cf_standard_name', default_list)
+
+        variables_list = []
+        for name, unit, long_name, standard_name in zip(variable_names, units, long_names, standard_names):
+            variables_list.append(dict(name=name, units=unit, long_name=long_name, standard_name=standard_name))
+
+        return variables_list
+
+    @property
+    def protocols(self) -> []:
+        if self._protocol_list is None:
+            self._protocol_list = [protocol for protocol in self._json_dict.get('access', [])
+                                   if protocol in _ODP_AVAILABLE_PROTOCOLS_LIST]
+        return self._protocol_list
+
+    def find_url(self, desired_service='HTTP'):
+        for url_service in self._json_dict.get('url', []):
+            parts = url_service.split('|')
+            if len(parts) == 2:
+                url, service = parts
+                if service == desired_service:
+                    return url
+        return None
+
     def update_file_list(self, monitor: Monitor = Monitor.NONE) -> None:
         self._file_list = None
-        asyncio.run(self._init_file_list(monitor))
+        self._init_file_list(monitor)
 
     def local_dataset_dir(self):
-        return os.path.join(get_data_store_path(), self._datasource_id)
+        return os.path.join(get_data_store_path(), self._master_id)
 
     def local_metadata_dataset_dir(self):
-        return os.path.join(get_metadata_store_path(), self._datasource_id)
+        return os.path.join(get_metadata_store_path(), self._master_id)
 
     def _find_files(self, time_range):
         requested_start_date, requested_end_date = time_range if time_range else (None, None)
-        asyncio.run(self._init_file_list())
+        self._init_file_list()
         if requested_start_date or requested_end_date:
             selected_file_list = []
             for file_rec in self._file_list:
@@ -1063,6 +853,13 @@ class EsaCciOdpOsDataSource(DataSource):
         :return:
         """
         return [file_rec[4][protocol].replace('.html', '') for file_rec in files_description_list]
+
+    @staticmethod
+    def _get_harmonized_coordinate_value(attrs: dict, attr_name: str):
+        value = attrs.get(attr_name, 'nan')
+        if isinstance(value, str):
+            return float(value.rstrip('degrees').rstrip('f'))
+        return value
 
     def _make_local(self,
                     local_ds: LocalDataSource,
@@ -1125,7 +922,7 @@ class EsaCciOdpOsDataSource(DataSource):
                             child_monitor.progress(work=20)
 
                             if var_names:
-                                remote_dataset = remote_dataset.drop_vars(
+                                remote_dataset = remote_dataset.drop(
                                     [var_name for var_name in remote_dataset.data_vars.keys()
                                      if var_name not in var_names]
                                 )
@@ -1261,6 +1058,7 @@ class EsaCciOdpOsDataSource(DataSource):
                     raise ValueError('Datastore {} already contains dataset {}'.format(local_store.id, ds_id))
 
         local_meta_info = self.meta_info.copy()
+        local_meta_info['ref_uuid'] = local_meta_info.get('uuid', None)
         local_meta_info['uuid'] = uuid
 
         local_ds = local_store.create_data_source(ds_id, title=title,
@@ -1287,30 +1085,12 @@ class EsaCciOdpOsDataSource(DataSource):
         else:
             return None
 
-    async def ensure_meta_info_set(self):
-        if self._meta_info_dict:
-            return
-        # todo set True when dimensions shall be read during meta data fetching
-        self._meta_info_dict = await _load_or_fetch_json(_fetch_meta_info,
-                                                   fetch_json_args=[self._raw_id,
-                                                                    self._json_dict['odd_url'],
-                                                                    self._json_dict['metadata_url'],
-                                                                    self._json_dict['variables'],
-                                                                    # True],
-                                                                    False],
-                                                   fetch_json_kwargs=dict(),
-                                                   cache_used=self._data_store.index_cache_used,
-                                                   cache_dir=self.local_metadata_dataset_dir(),
-                                                   cache_json_filename='meta-info.json',
-                                                   cache_timestamp_filename='meta-info-timestamp.txt',
-                                                   cache_expiration_days=self._data_store.index_cache_expiration_days)
-
-    async def _init_file_list(self, monitor: Monitor = Monitor.NONE):
-        await self.ensure_meta_info_set()
+    def _init_file_list(self, monitor: Monitor = Monitor.NONE):
         if self._file_list:
             return
-        file_list = await _load_or_fetch_json(_fetch_file_list_json,
-                                        fetch_json_args=[self._raw_id],
+
+        file_list = _load_or_fetch_json(_fetch_file_list_json,
+                                        fetch_json_args=[self._master_id, self._dataset_id],
                                         fetch_json_kwargs=dict(monitor=monitor),
                                         cache_used=self._data_store.index_cache_used,
                                         cache_dir=self.local_metadata_dataset_dir(),
@@ -1318,9 +1098,10 @@ class EsaCciOdpOsDataSource(DataSource):
                                         cache_timestamp_filename='file-list-timestamp.txt',
                                         cache_expiration_days=self._data_store.index_cache_expiration_days)
 
-        time_frequency = self._meta_info_dict.get('time_frequency', None)
-        if time_frequency is None and 'time_frequencies' in self._meta_info_dict:
-            time_frequency = self._meta_info_dict.get('time_frequencies')[0]
+        time_frequency = self._json_dict.get('time_frequency', None)
+        if time_frequency and isinstance(time_frequency, list):
+            time_frequency = time_frequency[0]
+
         if time_frequency:
             time_delta = _TIME_FREQUENCY_TO_TIME_DELTA.get(time_frequency, timedelta(days=0))
         else:
@@ -1333,7 +1114,7 @@ class EsaCciOdpOsDataSource(DataSource):
         # Compute the data source's temporal coverage
         for file_rec in file_list:
             if file_rec[1]:
-                file_start_date = datetime.strptime(file_rec[1].split('.')[0], _TIMESTAMP_FORMAT)
+                file_start_date = datetime.strptime(file_rec[1], _TIMESTAMP_FORMAT)
                 file_end_date = file_start_date + time_delta
                 data_source_start_date = min(data_source_start_date, file_start_date)
                 data_source_end_date = max(data_source_end_date, file_end_date)
@@ -1341,6 +1122,9 @@ class EsaCciOdpOsDataSource(DataSource):
                 file_rec[2] = file_end_date
         self._temporal_coverage = data_source_start_date, data_source_end_date
         self._file_list = file_list
+
+    def to_json(self):
+        return self._json_dict
 
     def __str__(self):
         return self.info_string
@@ -1374,3 +1158,91 @@ class _DownloadStatistics:
     @staticmethod
     def _to_megas(bytes_count: int) -> float:
         return bytes_count / (1000. * 1000.)
+
+
+class EsaCciCatalogueService:
+    def __init__(self, catalogue_url: str):
+
+        self._catalogue_url = catalogue_url
+
+        self._catalogue = None
+        self._catalogue_service = None
+
+        self._namespaces = Namespaces()
+
+    def reset(self):
+        self._catalogue = None
+        self._catalogue_service = None
+
+    def getrecords(self, monitor: Monitor = Monitor.NONE):
+        error_message = f"Failed accessing CCI ODP service {self._catalogue_url}"
+        try:
+            if not self._catalogue_service:
+                self._init_service()
+            if not self._catalogue:
+                self._build_catalogue(monitor.child(1))
+            return self._catalogue
+        except urllib.error.HTTPError as e:
+            raise DataAccessError(f"{error_message}: {e}") from e
+        except (urllib.error.URLError, socket.timeout) as e:
+            raise NetworkError(f"{error_message}: {e}") from e
+        except OSError as e:
+            raise DataAccessError(f"{error_message}: {e}") from e
+
+    def _build_catalogue(self, monitor: Monitor = Monitor.NONE):
+
+        self._catalogue = {}
+
+        catalogue_metadata = {}
+
+        start_position = 0
+        max_records = _CSW_MAX_RESULTS
+
+        matches = -1
+        while True:
+            # fetch record metadata
+            self._catalogue_service.getrecords2(esn='full', outputschema=self._namespaces.get_namespace('gmd'),
+                                                startposition=start_position, maxrecords=max_records)
+            if matches == -1:
+                # set counters, start progress monitor
+                matches = self._catalogue_service.results.get('matches')
+                if matches == 0:
+                    break
+                monitor.start(label="Fetching catalogue data... (%d records)" % matches,
+                              total_work=ceil(matches / max_records))
+
+            catalogue_metadata.update(self._catalogue_service.records)
+            monitor.progress(work=1)
+
+            # bump counters
+            start_position += max_records
+            if start_position > matches:
+                break
+
+        self._catalogue = {
+            record.identification.uricode[0]: {
+                'abstract': record.identification.abstract,
+                'bbox_minx': record.identification.bbox.minx if record.identification.bbox else None,
+                'bbox_miny': record.identification.bbox.miny if record.identification.bbox else None,
+                'bbox_maxx': record.identification.bbox.maxx if record.identification.bbox else None,
+                'bbox_maxy': record.identification.bbox.maxy if record.identification.bbox else None,
+                'creation_date':
+                    next(iter(e.date for e in record.identification.date if e and e.type == 'creation'), None),
+                'publication_date':
+                    next(iter(e.date for e in record.identification.date if e and e.type == 'publication'), None),
+                'title': record.identification.title,
+                'data_sources': record.identification.uricode[1:],
+                'licences': record.identification.uselimitation,
+                'temporal_coverage_start': record.identification.temporalextent_start,
+                'temporal_coverage_end': record.identification.temporalextent_end
+            }
+            for record in catalogue_metadata.values()
+            if record.identification and len(record.identification.uricode) > 0
+        }
+        monitor.done()
+
+    def _init_service(self):
+        if self._catalogue:
+            return
+        if not self._catalogue_service:
+            self._catalogue_service = CatalogueServiceWeb(url=self._catalogue_url, timeout=_CSW_TIMEOUT, skip_caps=True)
