+++ tests/ds/test_esa_cci_odp.py
@@ -1,17 +1,238 @@
-import datetime
+from datetime import datetime
+from lxml.etree import XML
+import asyncio
 import json
 import os
-import os.path
 import shutil
 import tempfile
 import unittest
 import unittest.mock
 import urllib.request
-
-from cate.core.ds import DATA_STORE_REGISTRY, DataAccessError, format_variables_info_string, DataStoreNotice
+from cate.core.ds import DATA_STORE_REGISTRY, DataAccessError, DataStoreNotice
+from cate.ds.esa_cci_odp import _fetch_file_list_json, _extract_metadata_from_odd, _extract_metadata_from_odd_url, \
+    _extract_metadata_from_descxml, _extract_metadata_from_descxml_url, _harmonize_info_field_names, \
+    _DownloadStatistics, EsaCciOdpDataStore, find_datetime_format, _retrieve_infos_from_dds
 from cate.core.types import PolygonLike, TimeRangeLike, VarNamesLike
-from cate.ds.esa_cci_odp import EsaCciOdpDataStore, find_datetime_format, _DownloadStatistics
 from cate.ds.local import LocalDataStore
+
+
+class EsaCciOdpOsTest(unittest.TestCase):
+
+    @unittest.skip(reason='Requires web access')
+    def test_fetch_opensearch_json_1(self):
+        file_list = asyncio.run(_fetch_file_list_json('49bcb6f29c824ae49e41d2d3656f11be'))
+        self.assertEqual("ESACCI-OC-L3S-K_490-MERGED-8D_DAILY_4km_GEO_PML_KD490_Lee-19970829-fv2.0.nc", file_list[0][0])
+        self.assertEqual("1997-09-04T00:00:00", file_list[0][1])
+        self.assertEqual("1997-09-04T00:00:00", file_list[0][2])
+        self.assertEqual(5924361, file_list[0][3])
+        self.assertEqual(2, len(file_list[0][4]))
+        self.assertTrue("Download" in file_list[0][4])
+        self.assertEqual("http://dap.ceda.ac.uk/thredds/fileServer/neodc/esacci/ocean_colour/data/v2-release/"
+                         "geographic/netcdf/kd/8day/v2.0/1997/"
+                         "ESACCI-OC-L3S-K_490-MERGED-8D_DAILY_4km_GEO_PML_KD490_Lee-19970829-fv2.0.nc",
+                         file_list[0][4].get("Download"))
+        self.assertEqual("http://dap.ceda.ac.uk/thredds/dodsC/dap//neodc/esacci/ocean_colour/data/v2-release/"
+                         "geographic/netcdf/kd/8day/v2.0/1997/"
+                         "ESACCI-OC-L3S-K_490-MERGED-8D_DAILY_4km_GEO_PML_KD490_Lee-19970829-fv2.0.nc",
+                         file_list[0][4].get("Opendap"))
+
+    @unittest.skip(reason='Requires web access. Also, we are getting an error 500 currently')
+    def test_fetch_opensearch_json_2(self):
+        file_list = asyncio.run(_fetch_file_list_json('b382ebe6679d44b8b0e68ea4ef4b701c'))
+        self.assertEqual(54, len(file_list))
+        self.assertEqual("ESACCI-LC-L4-LCCS-Map-300m-P1Y-1992-v2.0.7b.nc", file_list[1][0])
+        self.assertEqual("1992-01-01T00:00:00", file_list[1][1])
+        self.assertEqual("1992-01-01T00:00:00", file_list[1][2])
+        self.assertEqual(2543742078, file_list[1][3])
+        self.assertEqual(2, len(file_list[1][4]))
+        self.assertTrue("Download" in file_list[1][4])
+        self.assertEqual("http://cci-odp-data2.ceda.ac.uk/thredds/fileServer/esacci/land_cover/data/land_cover_maps/"
+                         "v2.0.7/ESACCI-LC-L4-LCCS-Map-300m-P1Y-1992-v2.0.7b.nc",
+                         file_list[1][4].get("Download"))
+        self.assertEqual('http://cci-odp-data2.ceda.ac.uk/thredds/dodsC/esacci/land_cover/data/land_cover_maps/'
+                         'v2.0.7/ESACCI-LC-L4-LCCS-Map-300m-P1Y-1992-v2.0.7b.nc',
+                         file_list[1][4].get("Opendap"))
+
+    def test_extract_metadata_from_odd_file(self):
+        odd_file = os.path.join(os.path.dirname(__file__), 'resources/odd.xml')
+        with open(odd_file) as odd:
+            json_obj = _extract_metadata_from_odd(XML(odd.read()))
+            self.assertFalse('query' in json_obj)
+            self.assertTrue('ecv' in json_obj)
+            self.assertEquals('LC', json_obj['ecv'])
+            self.assertTrue('time_frequencies' in json_obj)
+            self.assertEquals(['day', 'year'], json_obj['time_frequencies'])
+            self.assertTrue('institute' in json_obj)
+            self.assertEquals('Universite Catholique de Louvain', json_obj['institute'])
+            self.assertTrue('processing_level' in json_obj)
+            self.assertEquals('L4', json_obj['processing_level'])
+            self.assertTrue('product_string' in json_obj)
+            self.assertEquals('Map', json_obj['product_string'])
+            self.assertTrue('product_version' in json_obj)
+            self.assertEquals('2.0.7', json_obj['product_version'])
+            self.assertTrue('data_type' in json_obj)
+            self.assertEquals('LCCS', json_obj['data_type'])
+            self.assertFalse('sensor_id' in json_obj)
+            self.assertFalse('platform_id' in json_obj)
+            self.assertTrue('file_format' in json_obj)
+            self.assertEquals('.nc', json_obj['file_format'])
+
+    @unittest.skip(reason='Requires web access')
+    def test_extract_metadata_from_odd_url(self):
+        odd_url = 'http://opensearch-test.ceda.ac.uk/opensearch/description.xml?' \
+                  'parentIdentifier=7c7a38b2d2ce448b99194bff85a85248'
+        json_obj = _extract_metadata_from_odd_url(odd_url)
+        self.assertFalse('query' in json_obj)
+        self.assertTrue('ecv' in json_obj)
+        self.assertEquals('SOILMOISTURE', json_obj['ecv'])
+        self.assertTrue('time_frequency' in json_obj)
+        self.assertEquals('day', json_obj['time_frequency'])
+        self.assertFalse('institute' in json_obj)
+        self.assertTrue('processing_level' in json_obj)
+        self.assertEquals('L3S', json_obj['processing_level'])
+        self.assertTrue('product_string' in json_obj)
+        self.assertEquals('COMBINED', json_obj['product_string'])
+        self.assertTrue('product_version' in json_obj)
+        self.assertEquals('03.2', json_obj['product_version'])
+        self.assertTrue('data_type' in json_obj)
+        self.assertEquals('SSMV', json_obj['data_type'])
+        self.assertTrue('sensor_id' in json_obj)
+        self.assertEquals('multi-sensor', json_obj['sensor_id'])
+        self.assertTrue('platform_id' in json_obj)
+        self.assertEquals('multi-platform', json_obj['platform_id'])
+        self.assertTrue('file_format' in json_obj)
+        self.assertEquals('.nc', json_obj['file_format'])
+
+    @unittest.skip(reason='Requires web access')
+    def test_extract_metadata_from_descxml_url(self):
+        desc_url = 'https://catalogue.ceda.ac.uk/export/xml/49bcb6f29c824ae49e41d2d3656f11be.xml'
+        json_obj = asyncio.run(_extract_metadata_from_descxml_url(None, desc_url))
+        self.assert_json_obj_from_desc_xml(json_obj)
+
+    def test_extract_metadata_from_descxml(self):
+        desc_file = os.path.join(os.path.dirname(__file__), 'resources/49bcb6f29c824ae49e41d2d3656f11be.xml')
+        with open(desc_file) as desc:
+            json_obj = _extract_metadata_from_descxml(XML(desc.read()))
+            self.assert_json_obj_from_desc_xml(json_obj)
+
+    @unittest.skip(reason='Requires web access')
+    def test_retrieve_dimensions_from_dds_url(self):
+        dds_url = "http://dap.ceda.ac.uk/thredds/dodsC/dap//neodc/esacci/soil_moisture/data/daily_files/" \
+                  "COMBINED/v04.4/1986/ESACCI-SOILMOISTURE-L3S-SSMV-COMBINED-19861125000000-fv04.4.nc.dds"
+        response = urllib.request.urlopen(dds_url)
+        dimensions, variable_infos = _retrieve_infos_from_dds(response.readlines())
+        self.assertEquals(3, len(dimensions))
+        self.assertTrue('lat' in dimensions)
+        self.assertEquals(720, dimensions['lat'])
+        self.assertTrue('lon' in dimensions)
+        self.assertEquals(1440, dimensions['lon'])
+        self.assertTrue('time' in dimensions)
+        self.assertEquals(1, dimensions['time'])
+        self.assertEquals(11, len(variable_infos))
+        self.assertTrue('freqbandID' in variable_infos)
+        self.assertEquals('Int16', variable_infos['freqbandID']['data_type'])
+        self.assertEquals(3, len(variable_infos['freqbandID']['dimensions']))
+        self.assertTrue('time' in variable_infos['freqbandID']['dimensions'])
+        self.assertTrue('lat' in variable_infos['freqbandID']['dimensions'])
+        self.assertTrue('lon' in variable_infos['freqbandID']['dimensions'])
+
+    def test_retrieve_dimensions_from_dds(self):
+        dds_file = os.path.join(os.path.dirname(__file__),
+                                "resources/ESACCI-SOILMOISTURE-L3S-SSMV-COMBINED-19861125000000-fv04.4.nc.dds")
+        dds = open(dds_file)
+        dimensions, variable_infos = _retrieve_infos_from_dds(dds.readlines())
+        self.assertEquals(4, len(dimensions))
+        self.assertTrue('xc' in dimensions)
+        self.assertEquals(216, dimensions['xc'])
+        self.assertTrue('yc' in dimensions)
+        self.assertEquals(216, dimensions['yc'])
+        self.assertTrue('time' in dimensions)
+        self.assertEquals(1, dimensions['time'])
+        self.assertTrue('nv' in dimensions)
+        self.assertEquals(2, dimensions['nv'])
+        self.assertEquals(13, len(variable_infos))
+        self.assertTrue('Lambert_Azimuthal_Grid' in variable_infos)
+        self.assertEquals('Int32', variable_infos['Lambert_Azimuthal_Grid']['data_type'])
+        self.assertEquals(0, len(variable_infos['Lambert_Azimuthal_Grid']['dimensions']))
+        self.assertTrue('time_bnds' in variable_infos)
+        self.assertEquals('Float64', variable_infos['time_bnds']['data_type'])
+        self.assertEquals(2, len(variable_infos['time_bnds']['dimensions']))
+        self.assertTrue('time' in variable_infos['time_bnds']['dimensions'])
+        self.assertTrue('nv' in variable_infos['time_bnds']['dimensions'])
+        self.assertTrue('ice_conc' in variable_infos)
+        self.assertEquals('Int32', variable_infos['ice_conc']['data_type'])
+        self.assertEquals(3, len(variable_infos['ice_conc']['dimensions']))
+        self.assertTrue('time' in variable_infos['ice_conc']['dimensions'])
+        self.assertTrue('yc' in variable_infos['ice_conc']['dimensions'])
+        self.assertTrue('xc' in variable_infos['ice_conc']['dimensions'])
+
+    def assert_json_obj_from_desc_xml(self, json_obj: dict):
+        self.assertTrue('abstract' in json_obj)
+        self.maxDiff = None
+        self.assertTrue('title' in json_obj)
+        self.assertTrue('licences' in json_obj)
+        self.assertTrue('bbox_minx' in json_obj)
+        self.assertEquals('-180.0', json_obj['bbox_minx'])
+        self.assertTrue('bbox_miny' in json_obj)
+        self.assertEquals('-90.0', json_obj['bbox_miny'])
+        self.assertTrue('bbox_maxx' in json_obj)
+        self.assertEquals('180.0', json_obj['bbox_maxx'])
+        self.assertTrue('bbox_maxy' in json_obj)
+        self.assertEquals('90.0', json_obj['bbox_maxy'])
+        self.assertTrue('temporal_coverage_start' in json_obj)
+        self.assertEquals('1997-09-03T23:00:00', json_obj['temporal_coverage_start'])
+        self.assertTrue('temporal_coverage_end' in json_obj)
+        self.assertEquals('2013-12-31T23:59:59', json_obj['temporal_coverage_end'])
+        self.assertTrue('file_formats' in json_obj)
+        self.assertEquals('.nc', json_obj['file_formats'])
+        self.assertTrue('publication_date' in json_obj)
+        self.assertTrue('creation_date' in json_obj)
+        self.assertEquals('2016-12-12T17:08:42', json_obj['creation_date'])
+        self.assertTrue('platform_ids' in json_obj)
+        self.assertEquals(['Aqua', 'Envisat', 'Orbview-2'], json_obj['platform_ids'])
+        self.assertTrue('sensor_ids' in json_obj)
+        self.assertEquals(['MODIS', 'SeaWiFS', 'MERIS'], json_obj['sensor_ids'])
+        self.assertTrue('processing_levels' in json_obj)
+        self.assertEquals(['Level 3', 'Level 3S'], json_obj['processing_levels'])
+        self.assertTrue('time_frequencies' in json_obj)
+        self.assertEquals(['day', 'month', '8 days'], json_obj['time_frequencies'])
+
+    def test_harmonize_info_field_names(self):
+        test_catalogue = {'file_format': '.tiff', 'platform_ids': ['dfjh', 'ftrzg6'], 'sensor_id': 'hxfb75z',
+                          'sensor_ids': ['hxfb75z'], 'processing_level': 'L2', 'processing_levels': ['L1'],
+                          'time_frequency': 'gsyhdx', 'time_frequencies': ['gsyhdx', 'zsgsteczh', 'fzgu'],
+                          'field6': 'njgil', 'field6s': ['<dshjbre', 'hsr6u'], 'field7': 'multiple_field7s',
+                          'field7s': ['saedf', 'kihji']}
+
+        _harmonize_info_field_names(test_catalogue, 'file_format', 'file_formats')
+        _harmonize_info_field_names(test_catalogue, 'platform_id', 'platform_ids')
+        _harmonize_info_field_names(test_catalogue, 'sensor_id', 'sensor_ids')
+        _harmonize_info_field_names(test_catalogue, 'processing_level', 'processing_levels')
+        _harmonize_info_field_names(test_catalogue, 'time_frequency', 'time_frequencies')
+        _harmonize_info_field_names(test_catalogue, 'field6', 'field6s', 'multiple_field6s'),
+        _harmonize_info_field_names(test_catalogue, 'field7', 'field7s', 'multiple_field7s')
+
+        self.assertTrue('file_format' in test_catalogue)
+        self.assertEquals('.tiff', test_catalogue['file_format'])
+        self.assertFalse('file_formats' in test_catalogue)
+        self.assertFalse('platform_id' in test_catalogue)
+        self.assertTrue('platform_ids' in test_catalogue)
+        self.assertEquals(['dfjh', 'ftrzg6'], test_catalogue['platform_ids'])
+        self.assertTrue('sensor_id' in test_catalogue)
+        self.assertFalse('sensor_ids' in test_catalogue)
+        self.assertEquals('hxfb75z', test_catalogue['sensor_id'])
+        self.assertFalse('processing_level' in test_catalogue)
+        self.assertTrue('processing_levels' in test_catalogue)
+        self.assertEquals(['L1', 'L2'], test_catalogue['processing_levels'])
+        self.assertFalse('time_frequency' in test_catalogue)
+        self.assertTrue('time_frequencies' in test_catalogue)
+        self.assertEquals(['gsyhdx', 'zsgsteczh', 'fzgu'], test_catalogue['time_frequencies'])
+        self.assertFalse('field6' in test_catalogue)
+        self.assertTrue('field6s' in test_catalogue)
+        self.assertEquals(['<dshjbre', 'hsr6u', 'njgil'], test_catalogue['field6s'])
+        self.assertFalse('field7' in test_catalogue)
+        self.assertTrue('field7s' in test_catalogue)
+        self.assertEquals(['saedf', 'kihji'], test_catalogue['field7s'])
 
 
 @unittest.skip(reason='Because it writes a lot of files')
@@ -23,11 +244,10 @@ class EsaCciOdpDataStoreIndexCacheTest(unittest.TestCase):
         self.assertIsNotNone(data_sources)
         for data_source in data_sources:
             data_source.update_file_list()
-            # data_source.sync()
 
 
 def _create_test_data_store():
-    with open(os.path.join(os.path.dirname(__file__), 'esgf-index-cache.json')) as fp:
+    with open(os.path.join(os.path.dirname(__file__), 'resources/os-data-list.json')) as fp:
         json_text = fp.read()
     json_dict = json.loads(json_text)
     for d in DATA_STORE_REGISTRY.get_data_stores():
@@ -47,7 +267,7 @@ class EsaCciOdpDataStoreTest(unittest.TestCase):
 
     def test_id_title_and_is_local(self):
         self.assertEqual(self.data_store.id, 'test-odp')
-        self.assertEqual(self.data_store.title, 'ESA CCI Open Data Portal Legacy')
+        self.assertEqual(self.data_store.title, 'ESA CCI Open Data Portal')
         self.assertEqual(self.data_store.is_local, False)
 
     def test_description(self):
@@ -77,12 +297,12 @@ class EsaCciOdpDataStoreTest(unittest.TestCase):
     def test_query(self):
         data_sources = self.data_store.query()
         self.assertIsNotNone(data_sources)
-        self.assertEqual(len(data_sources), 61)
+        self.assertEqual(len(data_sources), 6)
 
     def test_query_with_string(self):
         data_sources = self.data_store.query(query_expr='OC')
         self.assertIsNotNone(data_sources)
-        self.assertEqual(len(data_sources), 20)
+        self.assertEqual(len(data_sources), 1)
 
 
 class EsaCciOdpDataSourceTest(unittest.TestCase):
@@ -104,10 +324,9 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
         self.data_store.get_updates(reset=True)
 
     def test_make_local_and_update(self):
-
-        soilmoisture_data_sources = self.data_store.query(
-            query_expr='esacci.SOILMOISTURE.day.L3S.SSMV.multi-sensor.multi-platform.COMBINED.02-1.r1')
-        soilmoisture_data_source = soilmoisture_data_sources[0]
+        soil_moisture_data_sources = self.data_store.query(
+            query_expr='esacci.20babc8f4dc449eaac11f47708e9f721')
+        soilmoisture_data_source = soil_moisture_data_sources[0]
 
         reference_path = os.path.join(os.path.dirname(__file__),
                                       os.path.normpath('resources/datasources/local/files/'))
@@ -117,23 +336,23 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
             def build_file_item(item_name: str, date_from: datetime, date_to: datetime, size: int):
 
                 return [item_name, date_from, date_to, size,
-                        {'OPENDAP': os.path.join(reference_path, item_name),
-                         'HTTPServer': 'file:' + urllib.request.pathname2url(os.path.join(reference_path, item_name))}]
+                        {'Opendap': os.path.join(reference_path, item_name),
+                         'Download': 'file:' + urllib.request.pathname2url(os.path.join(reference_path, item_name))}]
 
             reference_files = {
                 'ESACCI-SOILMOISTURE-L3S-SSMV-COMBINED-19781114000000-fv02.2.nc': {
-                    'date_from': datetime.datetime(1978, 11, 14, 0, 0),
-                    'date_to': datetime.datetime(1978, 11, 14, 23, 59),
+                    'date_from': datetime(1995, 11, 14, 0, 0),
+                    'date_to': datetime(1995, 11, 14, 23, 59),
                     'size': 21511378
                 },
                 'ESACCI-SOILMOISTURE-L3S-SSMV-COMBINED-19781115000000-fv02.2.nc': {
-                    'date_from': datetime.datetime(1978, 11, 15, 0, 0),
-                    'date_to': datetime.datetime(1978, 11, 15, 23, 59),
+                    'date_from': datetime(1995, 11, 15, 0, 0),
+                    'date_to': datetime(1995, 11, 15, 23, 59),
                     'size': 21511378
                 },
                 'ESACCI-SOILMOISTURE-L3S-SSMV-COMBINED-19781116000000-fv02.2.nc': {
-                    'date_from': datetime.datetime(1978, 11, 16, 0, 0),
-                    'date_to': datetime.datetime(1978, 11, 16, 23, 59),
+                    'date_from': datetime(1995, 11, 16, 0, 0),
+                    'date_to': datetime(1995, 11, 16, 23, 59),
                     'size': 21511378
                 }
             }
@@ -162,8 +381,10 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
             with unittest.mock.patch.object(EsaCciOdpDataStore, 'query', return_value=[]):
 
                 new_ds_title = 'local_ds_test'
-                new_ds_time_range = TimeRangeLike.convert((datetime.datetime(1978, 11, 14, 0, 0),
-                                                           datetime.datetime(1978, 11, 16, 23, 59)))
+                new_ds_time_range = TimeRangeLike.convert((datetime(1995, 11, 14, 0, 0),
+                                                           datetime(1995, 11, 16, 23, 59)))
+                # new_ds_time_range = TimeRangeLike.convert((datetime(1997, 5, 10, 0, 0),
+                #                                            datetime(1997, 5, 12, 23, 59)))
                 try:
                     new_ds = soilmoisture_data_source.make_local(new_ds_title, time_range=new_ds_time_range)
                 except Exception:
@@ -174,8 +395,8 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
                 self.assertEqual(new_ds.temporal_coverage(), new_ds_time_range)
 
                 new_ds_w_one_variable_title = 'local_ds_test_var'
-                new_ds_w_one_variable_time_range = TimeRangeLike.convert((datetime.datetime(1978, 11, 14, 0, 0),
-                                                                          datetime.datetime(1978, 11, 16, 23, 59)))
+                new_ds_w_one_variable_time_range = TimeRangeLike.convert((datetime(1995, 11, 14, 0, 0),
+                                                                          datetime(1995, 11, 16, 23, 59)))
                 new_ds_w_one_variable_var_names = VarNamesLike.convert(['sm'])
 
                 new_ds_w_one_variable = soilmoisture_data_source.make_local(
@@ -194,8 +415,8 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
                                     set(new_ds_w_one_variable_var_names))
 
                 new_ds_w_region_title = 'from_local_to_local_region'
-                new_ds_w_region_time_range = TimeRangeLike.convert((datetime.datetime(1978, 11, 14, 0, 0),
-                                                                    datetime.datetime(1978, 11, 16, 23, 59)))
+                new_ds_w_region_time_range = TimeRangeLike.convert((datetime(1995, 11, 14, 0, 0),
+                                                                    datetime(1995, 11, 16, 23, 59)))
                 new_ds_w_region_spatial_coverage = PolygonLike.convert("10,20,30,40")
 
                 new_ds_w_region = soilmoisture_data_source.make_local(
@@ -210,8 +431,8 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
                 self.assertEqual(new_ds_w_region.spatial_coverage(), new_ds_w_region_spatial_coverage)
 
                 new_ds_w_region_title = 'from_local_to_local_region_one_var'
-                new_ds_w_region_time_range = TimeRangeLike.convert((datetime.datetime(1978, 11, 14, 0, 0),
-                                                                    datetime.datetime(1978, 11, 16, 23, 59)))
+                new_ds_w_region_time_range = TimeRangeLike.convert((datetime(1995, 11, 14, 0, 0),
+                                                                    datetime(1995, 11, 16, 23, 59)))
                 new_ds_w_region_var_names = VarNamesLike.convert(['sm'])
                 new_ds_w_region_spatial_coverage = PolygonLike.convert("10,20,30,40")
 
@@ -232,8 +453,8 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
                 self.assertSetEqual(set(data_set.variables), set(new_ds_w_region_var_names))
 
                 new_ds_w_region_title = 'from_local_to_local_region_two_var_sm_uncertainty'
-                new_ds_w_region_time_range = TimeRangeLike.convert((datetime.datetime(1978, 11, 14, 0, 0),
-                                                                    datetime.datetime(1978, 11, 16, 23, 59)))
+                new_ds_w_region_time_range = TimeRangeLike.convert((datetime(1995, 11, 14, 0, 0),
+                                                                    datetime(1995, 11, 16, 23, 59)))
                 new_ds_w_region_var_names = VarNamesLike.convert(['sm', 'sm_uncertainty'])
                 new_ds_w_region_spatial_coverage = PolygonLike.convert("10,20,30,40")
 
@@ -253,7 +474,7 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
 
                 self.assertSetEqual(set(data_set.variables), set(new_ds_w_region_var_names))
 
-                empty_ds_timerange = (datetime.datetime(2017, 12, 1, 0, 0), datetime.datetime(2017, 12, 31, 23, 59))
+                empty_ds_timerange = (datetime(1917, 12, 1, 0, 0), datetime(1917, 12, 31, 23, 59))
                 with self.assertRaises(DataAccessError) as cm:
                     soilmoisture_data_source.make_local('empty_ds', time_range=empty_ds_timerange)
                 self.assertEqual(f'Data source "{soilmoisture_data_source.id}" does not'
@@ -261,8 +482,8 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
                                  f' time range {TimeRangeLike.format(empty_ds_timerange)}',
                                  str(cm.exception))
 
-                new_ds_time_range = TimeRangeLike.convert((datetime.datetime(1978, 11, 14, 0, 0),
-                                                           datetime.datetime(1978, 11, 14, 23, 59)))
+                new_ds_time_range = TimeRangeLike.convert((datetime(1995, 11, 14, 0, 0),
+                                                           datetime(1995, 11, 14, 23, 59)))
 
                 new_ds = soilmoisture_data_source.make_local("title_test_copy", time_range=new_ds_time_range)
                 self.assertIsNotNone(new_ds)
@@ -278,29 +499,15 @@ class EsaCciOdpDataSourceTest(unittest.TestCase):
                       self.data_store)
 
     def test_id(self):
-        self.assertEqual(self.first_oc_data_source.id,
-                         'esacci.OC.day.L3S.K_490.multi-sensor.multi-platform.MERGED.1-0.r2')
+        self.assertEqual(self.first_oc_data_source.id, 'esacci.915d2340b178494f987a6942e263a2eb')
 
     def test_schema(self):
         self.assertEqual(self.first_oc_data_source.schema,
                          None)
 
-    @unittest.skip(reason='outdated info string')
-    def test_info_string(self):
-        self.assertIn('product_string:          MERGED\n',
-                      self.first_oc_data_source.info_string)
-
-    def test_variables_info_string(self):
-        self.assertIn('kd_490 (m-1):\n',
-                      format_variables_info_string(self.first_oc_data_source.variables_info),
-                      self.first_oc_data_source.variables_info)
-        self.assertIn('Long name:        Downwelling attenuation coefficient at 490nm',
-                      format_variables_info_string(self.first_oc_data_source.variables_info))
-
-    @unittest.skip(reason='ssl error on windows')
     def test_temporal_coverage(self):
         self.assertEqual(self.first_oc_data_source.temporal_coverage(),
-                         (datetime.datetime(1997, 9, 4, 0, 0), datetime.datetime(2000, 6, 24, 0, 0)))
+                         (datetime(1997, 9, 3, 23, 0, 0), datetime(2016, 12, 31, 23, 59, 59)))
 
     def assert_tf(self, filename: str, expected_time_format: str):
         time_format, p1, p2 = find_datetime_format(filename)
@@ -386,12 +593,13 @@ class DownloadStatisticsTest(unittest.TestCase):
 @unittest.skip(reason='Used for debugging to fix Cate issues #823, #822, #818, #816, #783')
 class SpatialSubsetTest(unittest.TestCase):
 
+    @unittest.skip(reason='Requires variable access which is not integrated yet.')
     def test_make_local_spatial(self):
         data_store = EsaCciOdpDataStore()
-        data_source = data_store.query(ds_id='esacci.SST.day.L4.SSTdepth.multi-sensor.multi-platform.OSTIA.1-1.r1')[0]
+        data_source = data_store.query(ds_id='esacci.5db2099606b94e63879d841c87e654ae')[0]
         # The following always worked fine:
-        ds = data_source.open_dataset(time_range=['2010-01-01', '2010-01-04'], region='-10,40,20,70')
+        ds = data_source.open_dataset(time_range=['2004-01-01', '2004-01-14'], region='-10,40,20,70')
         self.assertIsNotNone(ds)
         # The following reproduced Cate issues #823, #822, #818, #816, #783:
-        ds = data_source.make_local('SST_DAY_L4', time_range=['2010-01-01', '2010-01-04'], region='-10,40,20,70')
+        ds = data_source.make_local('local_name', time_range=['2004-01-01', '2004-01-14'], region='-10,40,20,70')
         self.assertIsNotNone(ds)
